refactor(block): unify journal eviction + GC segment reclamation
Fold evict.go and gc.go into one reclaim.go and extract the shared
retireSegment primitive that eviction and GC both end on: drop from the
sealed set under sh.mu, drain in-flight preads via readGuard, close,
unlink .seg/.idx, and decrement diskBytes. The retirement tail was
duplicated ~3x (evictSegment, repackSegment step 5, dropVictim) and had
drifted.

Fixes two inconsistencies the duplication hid:
- readGuard drain: GC drained in-flight preads before close; eviction did
  not. Eviction now drains too (strictly safer, no behavior regression).
- diskBytes: GC never decremented diskBytes for the reclaimed victim and
  never counted the target's relocated records, so it drifted on both
  sides. retireSegment now decrements the victim; repackSegment adds the
  target's durable record bytes.

The two policies stay separate; only the final drop+drain+close+unlink+
accounting step is shared.
