+++ pkg/block/journal/reclaim.go
@@ -3,9 +3,222 @@ package journal
 import (
 	"context"
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"os"
+	"time"
+
+	"github.com/marmos91/dittofs/internal/logger"
 )
+
+// Segment reclamation: two distinct policies over one shared retirement tail.
+//
+//   - Eviction (disk pressure) marks a sealed, fully-synced segment's intervals
+//     cold — its bytes now live only remotely, reads refetch — then retires it.
+//   - GC/repack (dead-byte ratio) relocates a victim's still-live records into a
+//     fresh local segment and repoints the index — data stays warm — then retires
+//     the victim.
+//
+// Both end by dropping the segment from the sealed set, draining in-flight preads,
+// closing the fd, unlinking the files, and decrementing diskBytes: retireSegment.
+
+// ErrLocalStoreFull is returned by the write path when the local cache is at its
+// MaxLocalBytes cap and no segment can be evicted because every one is pinned by
+// unsynced (dirty) records. The writer must retry after carve drains those bytes
+// to the remote store; UnsyncedBytes reports how much is pinned.
+var ErrLocalStoreFull = errors.New("journal: local store full, all segments pinned by unsynced bytes")
+
+// evictBackoff is the poll interval the write path waits between eviction
+// attempts while dirty-pinned. The tunable knobs are Config.MaxLocalBytes (the
+// threshold) and Config.EvictMaxWait (the give-up budget); this fixed step just
+// paces the wait and is not worth a config field.
+const evictBackoff = 10 * time.Millisecond
+
+// EvictResult reports what an eviction pass reclaimed.
+type EvictResult struct {
+	SegmentsEvicted int
+	BytesFreed      int64
+}
+
+// Evict frees whole sealed segments under storage pressure, coldest first
+// (approx-LRU by lastAccess). Only sealed, fully-synced segments qualify: a
+// segment holding any unsynced record is skipped so eviction never destroys the
+// only copy of dirty bytes (the synced-gate, mirroring logblob.EvictBlob). It
+// evicts until targetBytes have been freed; targetBytes <= 0 evicts a single
+// qualifying segment. It returns what it reclaimed — SegmentsEvicted == 0 means
+// nothing qualified (the caller should backpressure, see ensureSpace).
+func (s *Store) Evict(ctx context.Context, targetBytes int64) (EvictResult, error) {
+	if err := ctx.Err(); err != nil {
+		return EvictResult{}, err
+	}
+	if s.closed.Load() {
+		return EvictResult{}, errClosed
+	}
+	var res EvictResult
+	for {
+		seg, sh := s.claimColdestEvictable()
+		if seg == nil {
+			return res, nil
+		}
+		freed, err := s.evictSegment(sh, seg)
+		if err != nil {
+			return res, err
+		}
+		res.SegmentsEvicted++
+		res.BytesFreed += freed
+		if targetBytes <= 0 || res.BytesFreed >= targetBytes {
+			return res, nil
+		}
+	}
+}
+
+// claimColdestEvictable finds the coldest sealed, fully-synced, unclaimed segment
+// across all shards and claims it (busy CAS) so eviction and GC never race on the
+// same segment. It returns nil only when no segment qualifies. Losing the claim
+// to a concurrent GC/evict does not abort the pass: that segment is now busy and
+// drops out of the next scan, so the coldest remaining candidate is tried
+// instead. The retry terminates because a lost CAS means the segment is busy,
+// shrinking the candidate set each round.
+func (s *Store) claimColdestEvictable() (*segmentMeta, *shard) {
+	for {
+		var (
+			best       *segmentMeta
+			bestShard  *shard
+			bestAccess int64
+		)
+		for _, sh := range s.shards {
+			sh.mu.Lock()
+			for _, seg := range sh.sealed {
+				if !evictable(seg) {
+					continue
+				}
+				if la := seg.lastAccess.Load(); best == nil || la < bestAccess {
+					best, bestShard, bestAccess = seg, sh, la
+				}
+			}
+			sh.mu.Unlock()
+		}
+		if best == nil {
+			return nil, nil
+		}
+		// The synced-gate only ever loosens for a sealed segment (carve raises
+		// syncedRecords toward records; records is frozen once sealed), so the sole
+		// concurrency hazard is another claimer — the CAS settles it.
+		if best.busy.CompareAndSwap(false, true) {
+			return best, bestShard
+		}
+	}
+}
+
+// evictable reports whether seg can be dropped whole: sealed, unclaimed, and with
+// every record synced to the remote store. Caller holds seg's shard lock.
+func evictable(seg *segmentMeta) bool {
+	return seg.sealed.Load() && !seg.busy.Load() &&
+		seg.syncedRecords.Load() == seg.records.Load()
+}
+
+// evictSegment removes one claimed sealed segment. Under the shard lock it
+// rewrites every interval-tree entry backed by the segment to a cold marker — so
+// a later read of that range fetches from the remote store instead of seeing a
+// false POSIX hole, and DataExtents still reports the range as present — then
+// retires the segment. Caller must have claimed seg.busy.
+func (s *Store) evictSegment(sh *shard, seg *segmentMeta) (int64, error) {
+	sh.mu.Lock()
+	for _, fi := range sh.index {
+		for k := range fi.ivs {
+			if fi.ivs[k].loc.SegmentID == seg.id && !fi.ivs[k].cold {
+				fi.ivs[k].cold = true
+			}
+		}
+	}
+	sh.mu.Unlock()
+	return s.retireSegment(sh, seg)
+}
+
+// ensureSpace is the write-path capacity gate. With MaxLocalBytes set, it evicts
+// cold synced segments to fit the incoming write; when nothing is evictable
+// because every segment is dirty-pinned, it backpressures the writer up to
+// EvictMaxWait (giving carve time to drain to the remote) and finally returns
+// ErrLocalStoreFull. A no-op when MaxLocalBytes is unset. Holds no lock, so the
+// eviction it drives never contends with the caller's own shard.
+//
+// MaxLocalBytes is a soft pressure threshold, not a hard byte quota: admission
+// reads diskBytes without reserving, so N concurrent writers across shards can
+// each clear the gate and append before any lands, briefly overshooting the cap
+// — and eviction is whole-segment anyway, so exact enforcement is neither
+// possible nor the goal. The gate relieves pressure (evict) and, failing that,
+// backpressures; a later writer's round evicts the overshoot. ErrLocalStoreFull
+// means genuinely nothing is evictable, never mere overshoot. A hard ceiling
+// would need a global reserved-bytes counter, deliberately not built (the
+// eviction design is lazy and pressure-gated).
+func (s *Store) ensureSpace(ctx context.Context, needed int64) error {
+	if s.cfg.MaxLocalBytes <= 0 {
+		return nil
+	}
+	deadline := time.Now().Add(s.cfg.EvictMaxWait)
+	warned := false
+	for s.diskBytes.Load()+needed > s.cfg.MaxLocalBytes {
+		if err := ctx.Err(); err != nil {
+			return err
+		}
+		overage := s.diskBytes.Load() + needed - s.cfg.MaxLocalBytes
+		res, err := s.Evict(ctx, overage)
+		if err != nil {
+			return err
+		}
+		if res.SegmentsEvicted > 0 {
+			continue
+		}
+		if !warned {
+			logger.Warn("journal local store full: every segment pinned by unsynced bytes, backpressuring writes until carve drains to remote",
+				"dir", s.dir,
+				"disk_bytes", s.diskBytes.Load(),
+				"max_local_bytes", s.cfg.MaxLocalBytes,
+				"unsynced_bytes", s.unsynced.Load())
+			warned = true
+		}
+		if time.Now().After(deadline) {
+			return ErrLocalStoreFull
+		}
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-time.After(evictBackoff):
+		}
+	}
+	return nil
+}
+
+// retireSegment is the shared reclamation tail both eviction and GC end on. The
+// caller has already done its policy-specific index work (evict: mark intervals
+// cold; GC: repoint the index to the relocation target) and claimed seg.busy.
+// This drops the segment from the sealed set under sh.mu, drains in-flight
+// unlocked preads via the exclusive readGuard — taken WITHOUT sh.mu so it can't
+// deadlock a reader that holds sh.mu while acquiring the shared guard — closes
+// the fd, unlinks the .seg/.idx files, and decrements diskBytes by the reclaimed
+// on-disk bytes. It returns those bytes.
+//
+// On unlink failure the segment is already out of the index and its fd closed —
+// reclaim is one-way. The leftover .seg is a harmless orphan the recovery sweep
+// reclaims; report the error and leave diskBytes counting it, since it still
+// occupies disk.
+func (s *Store) retireSegment(sh *shard, seg *segmentMeta) (int64, error) {
+	sh.mu.Lock()
+	freed := seg.tail.Load()
+	delete(sh.sealed, seg.id)
+	sh.mu.Unlock()
+
+	seg.readGuard.Lock()
+	_ = seg.close()
+	seg.readGuard.Unlock()
+
+	if err := os.Remove(s.segPath(seg.id)); err != nil && !errors.Is(err, os.ErrNotExist) {
+		return 0, fmt.Errorf("journal: retire remove segment %d: %w", seg.id, err)
+	}
+	_ = os.Remove(s.idxPath(seg.id)) // best-effort: the .idx is rebuildable
+	s.diskBytes.Add(-freed)
+	return freed, nil
+}
 
 // Garbage collection / repack.
 //
@@ -189,14 +402,14 @@ func findMove(moves []liveRec, lo, hi int64, ver uint64) int {
 }
 
 // repackSegment relocates a victim's still-live records (and carries its
-// tombstones forward) into a fresh sealed target, then unlinks the victim.
+// tombstones forward) into a fresh sealed target, then retires the victim.
 //
 // Ordering invariant — bytes-durable-before-index-durable-before-reclaim:
 //  1. write survivors into the target;
 //  2. fsync + seal the target (its dir entry was fsynced at create) — bytes durable;
 //  3. repoint the in-memory index to the target (the index is not itself durable
 //     — recovery rebuilds it from the .seg records, which are now durable);
-//  4. only then unlink the victim.
+//  4. only then retire the victim.
 //
 // A crash before step 4 leaves the victim as a harmless orphan: its records
 // duplicate the target's with the identical Version, so recovery replays both
@@ -277,6 +490,10 @@ func (s *Store) repackSegment(sh *shard, victim *segmentMeta) (int64, error) {
 	}
 	target.liveBytes.Store(relocated)
 	target.syncedRecords.Store(syncedCount)
+	// Account the relocated records + tombstones now durable in the target;
+	// writeDataRecord/writeTombstoneRecord don't touch diskBytes (unlike the
+	// append path), and createSegment already counted segHeaderSize.
+	s.diskBytes.Add(target.tail.Load() - segHeaderSize)
 
 	// 4. Repoint the index, skipping intervals a concurrent write superseded.
 	sh.mu.Lock()
@@ -319,18 +536,10 @@ func (s *Store) repackSegment(sh *shard, victim *segmentMeta) (int64, error) {
 		return 0, nil
 	}
 
-	// 5. Reclaim: drop the victim from the set so no new read resolves it, drain
-	// in-flight preads via the exclusive guard, then close and unlink. The guard
-	// is taken WITHOUT holding sh.mu so it can never deadlock against a reader
-	// that holds sh.mu while acquiring the shared guard.
-	sh.mu.Lock()
-	delete(sh.sealed, victim.id)
-	sh.mu.Unlock()
-	victim.readGuard.Lock()
-	_ = victim.close()
-	victim.readGuard.Unlock()
-	_ = os.Remove(s.segPath(victim.id))
-	_ = os.Remove(s.idxPath(victim.id))
+	// 5. Retire the victim: drop it, drain in-flight preads, close, unlink.
+	if _, err := s.retireSegment(sh, victim); err != nil {
+		return 0, err
+	}
 
 	net := occupied - relocated
 	if net < 0 {
@@ -345,7 +554,7 @@ func (s *Store) repackSegment(sh *shard, victim *segmentMeta) (int64, error) {
 var testStopBeforeUnlink bool
 
 // dropVictim reclaims a fully-dead segment: no live records, no tombstones. It
-// re-checks under the shard lock that nothing references it, then unlinks.
+// re-checks under the shard lock that nothing references it, then retires it.
 func (s *Store) dropVictim(sh *shard, victim *segmentMeta, occupied int64) (int64, error) {
 	sh.mu.Lock()
 	for _, fi := range sh.index {
@@ -356,13 +565,10 @@ func (s *Store) dropVictim(sh *shard, victim *segmentMeta, occupied int64) (int6
 			}
 		}
 	}
-	delete(sh.sealed, victim.id)
 	sh.mu.Unlock()
-	victim.readGuard.Lock()
-	_ = victim.close()
-	victim.readGuard.Unlock()
-	_ = os.Remove(s.segPath(victim.id))
-	_ = os.Remove(s.idxPath(victim.id))
+	if _, err := s.retireSegment(sh, victim); err != nil {
+		return 0, err
+	}
 	return occupied, nil
 }
